+++ CHANGELOG.md
@@ -126,3 +126,4 @@ posts which too old or too young
 
 * Now you can subscribe directly on `ReceiveChannel` updates
 * Rewrite `debounce`, now it is more compact and effective
+* Proxy settings was replaced into `clientConfig` settings